+++ ProjectWings/Source/ProjectWings/Public/Core/WingsPlayerController.h
@@ -23,5 +23,8 @@ protected:
 public:
     /** 발사 시 기체로 조종권을 넘기는 함수 */
     void TransitionToFlight(APawn* FlightPawn, APawn* PreviousPawn = nullptr);
+
+    /** 발사대로 조종권을 다시 돌리는 함수 (재시도용) */
+    void ReturnToLauncher();
 	
 };